requirements.txt
Flask==2.0.1
Flask_Admin==1.5.8
WTForms==2.3.3
itsdangerous==2.0.1
Flask_Login==0.5.0
SQLAlchemy==1.4.20
Flask_Bcrypt==0.7.1
Flask_WTF==0.15.1
Flask_SQLAlchemy==2.5.1
Flask_Mail==0.9.1
python_dateutil==2.8.1
